Rename several tag files to one
