+++ work/templates/work/compact_search_table_layout.html
@@ -1,4 +1,4 @@
-{% load cl_filters render_queryable_resources render_queryable_images render_queryable_manif %}
+{% load cl_filters work_util_tags %}
 
 <script>
     let tippies = [];
