+++ work/templates/work/component/exclamation.html
@@ -1,4 +1,4 @@
-{% load exclamation %}
+{% load work_util_tags %}
 
 {% if work|exclamation != '' %}
